intcode: Slight refactor for day09, not quiet there yet.

+++ aoc2019/src/day02/mod.rs
@@ -1,7 +1,7 @@
 #[path = "../intcode/mod.rs"]
 mod intcode;
 
-pub fn run_intcode(insts: &[i32]) -> Vec<i32> {
+pub fn run_intcode(insts: &[i64]) -> Vec<i64> {
   let mut program = intcode::Intcode::new(insts, &[]);
   program.run();
   program.insts
@@ -11,14 +11,14 @@ pub fn raw_input() -> String {
   include_str!("input.txt").to_string()
 }
 
-pub fn parse_input(input: &str) -> Vec<i32> {
+pub fn parse_input(input: &str) -> Vec<i64> {
   input
     .split(',')
     .map(|e| e.replace(",", "").replace("+", "").parse::<_>().unwrap())
     .collect()
 }
 
-fn get_registry(insts: &mut Vec<i32>, operation: i32, pc: usize, param: usize, addr: bool) -> i32 {
+fn get_registry(insts: &mut Vec<i64>, operation: i64, pc: usize, param: usize, addr: bool) -> i64 {
   let val = insts[pc + param];
   let mask = (operation as u32) / (10_u32.pow(param as u32 + 1)) % 10;
   assert!(!addr || mask == 0);
@@ -29,14 +29,14 @@ fn get_registry(insts: &mut Vec<i32>, operation: i32, pc: usize, param: usize, a
   }
 }
 
-pub fn part1(input: &[i32]) -> i32 {
+pub fn part1(input: &[i64]) -> i64 {
   let mut insts = input.to_vec();
   insts[1] = 12;
   insts[2] = 2;
   run_intcode(&insts)[0]
 }
 
-pub fn part2(input: &[i32]) -> i32 {
+pub fn part2(input: &[i64]) -> i64 {
   for noun in 0..100 {
     for verb in 0..100 {
       let mut insts = input.to_vec();
